Updating recipe or plan does not show updated instructions
The instructions markup is generated by the server and is displayed in
the recipe display view. The display view assumes that the recipe has
been processed by the server and that the appropriate HTML markup has
been generated. The server does return the updated document from a
successful call to PUT or POST but the API was not returning that data
to the view.

The API has been updated to return the updated document for plans,
recipes, and shopping lists. The views for plans and recipes make use
of this data in order to properly show the real state of the document.