My Shopping List tab is available when not logged in
The set of navigation tabs was hardcoded into the HTML despite not all
tabs being relevant for users that have not signed in. This adds
support for dynamically adding and removing tabs and converts the My
Shopping List tab to a dynamic tab that is available only once the
user logs in.