Site does not work in IE11
IE11 takes issue with es6-shim after 0.24.0. It throws access denied
errors when the shim polyfills getOwnPropertyNames for primatives. As
a result, revert to 0.24.0 of es6-shim since it is the last release
that IE11 can use.

Further, the refactoring began to use Symbols, which IE does not
support. Reference es6-symbol as a polyfill.

IE10 still fails because it has only a partial implementation of the
dataset attribute.