New button displays even when viewing the shopping list
The button to create a recipe or meal plan remained visible even when
viewing the shopping list because it was hardcoded as a member of the
secondary menu. This created confusing UI since the button has no
meaning in the shopping list view.

The button is no longer hard coded and is instead prepended to the
list of tags that are generated for meal plans and recipes.