After selecting a tag it is difficult to return to the list of tags
Several solutions were thought of to solve this:

* Adding a back button to the listings which takes you back to the
  tags

* Using bread crumbs

* Having any click on the main navigation button return to tags view

* Displaying the list of items in the content area

The final solution was chosen because it is the simplest UI and it
matches the conventions of other tag and category-based navigation
schemes. With this change the side menu now always is a set of tags,
further simplifying the mental model of what the menu represents.

Since the list of items in the content draws an opaque box around each
item the fact that the content area doesn't span the rest of the width
of the page became apparent. To fix this the tag menu is made to float
and the content area uses `overflow: hidden` to trigger CSS block
formatting context, causing it to take up all remaining width.