Add support for tracking ideas
Each user may now have some set of ideas that they want to keep track
of. These ideas may be related to thoughts about recipes, ingredients,
or tecnniques that they wish to experiment with, for example. Each
idea has a description (the text of the idea) as well as an optional
source or inspiration of the idea.

Ideas are per-user and are surfaced once the user logs in. They may be
added one at a time. A user can either view all ideas at once or may
view a random five.

Internally, ideas are stored as a simple list of objects containing a
string description and an optional source string. They are not
relational and so the entire set of ideas is stored in a single row.

Currently there is no limit to the number of ideas that a user may
have or the size of each individual idea.

A few style issues were fixed as a result of this change. In
particular, the top-level menu items were separated not by a proper
margin between elements but by a text node that existed between each
`li` element within the list. The text node has been removed and the
margin style is now used to properly control the separation of the
menu items.